strsignal(INT_MIN) reads off the front of a string literal, and a lint for the shape
Three defects in one day were one shape: systat's get_tbl_ptr(), libc's
querylocale() and libcalendar's easterodn(). The first two write a bound
for a two-sided index as one comparison -- and the first says so out
loud in its own comment, "if our index is out of range", while the code
checks one end. The third subscripts with %, which in C keeps the sign
of the dividend.

tools/verify/onesided_index.py looks for both, and found a fourth.

	signum = num;
	if (num < 0)
		signum = -signum;
	...
	*t++ = "0123456789"[signum % 10];

strsignal(3) takes a plain int and that else arm handles EVERY value
outside the signal range, so strsignal(INT_MIN) reaches it. Negating the
most negative int is undefined, and it does not even come out positive:
signum stays negative and the subscript reads off the front of the
string literal, into the buffer strsignal() returns. The file had
carried `XXX: negative num ? (REGR)' above the definition. Somebody
knew.

strerror()'s errstr() had the same line -- uerr = (num >= 0) ? num :
-num -- and though the destination was already unsigned, -num is
computed in int before it gets there. Same fix, negate in unsigned,
which is exact and modular. Rendered over the full range afterwards,
INT_MIN included, and every value is unchanged.

The ANALYSER has nothing here: lib/libc/string measures 0 findings
across 86 translation units. Neither does the model checker. This one
came from the lint alone.

ONLY ONE OF THE TWO RULES GATES

MOD reports 8 sites tree-wide, every one read and written into EXPECTED
with the reason: a fraction-digit count from strlen, pr's own line
counter, two pid_t from fork(2), a queue rank, a ring index counted from
zero, a compat_strtoul() result, an unsigned buffer offset. That is a
gate, and it bites -- removing one record fails it, which is a case in
the test file rather than a claim.

ONESIDED reports 264, down from 1,109 over four rounds, and they are
dominated by internal contracts. A gate demanding 264 readings before
the next commit is a gate somebody turns off, so it prints and does not
fail -- the same call noreturn_check.py makes about its 673.

Each tightening round was the checker being wrong about the tree: a
`for (V = ...)' initialiser is a loop counter whatever the initialiser
is (1109 -> 656); assert(0 < i && ...) is the author stating the bound
(-> 569); an index whose value never leaves the function is not this
defect (-> 264); one report per (function, variable), not per use. And
two rounds before that were the checker failing to find the defects it
was written for -- the systat shape puts its ceiling on the SOURCE
expression, and `if ((scale = get_scale(args)) != -1)' puts the floor
and the name on either side of a call. All of them are cases now, along
with the three originals in the form they had before they were fixed.

+++ hbsd/src/lib/libc/string/strerror.c
@@ -64,7 +64,13 @@ errstr(int num, const char *uprefix, char *buf, size_t len)
 
 	t = tmp + sizeof(tmp);
 	*--t = '\0';
-	uerr = (num >= 0) ? num : -num;
+	/*
+	 * The negation in UNSIGNED: `-num' is computed in int before it
+	 * reaches uerr, so errstr(INT_MIN, ...) was undefined even
+	 * though the destination could hold the answer.  strsignal()
+	 * had the same line with the same fix.
+	 */
+	uerr = (num >= 0) ? (unsigned int)num : -(unsigned int)num;
 	do {
 		*--t = "0123456789"[uerr % 10];
 	} while (uerr /= 10);
